Move tests to integration folder
